+++ ScannerMonitor/HWGrowlScannerMonitor.h
@@ -1,0 +1,11 @@
+//
+//  HWGrowlScannerMonitor.h
+//  HardwareGrowler
+//
+
+#import <Foundation/Foundation.h>
+#import "HardwareGrowlPlugin.h"
+
+@interface HWGrowlScannerMonitor : NSObject <HWGrowlPluginProtocol, HWGrowlPluginNotifierProtocol>
+
+@end